Fix improper argument serialization on ARGV passing tecnique
ciscocmd uses a tecnique to pass arguments to expect by means of
serialization. The tecnique incorrectly used IFS combined with $@
instead of "$*".

This worked on Bash 4.3 because of a bug that was fixed on 4.4,
which breaks the tecnique.

This fixes the usage and restores ciscocmd functionality on Bash 4.4.

This fixes issue #2 on Github:
* https://github.com/eczema/ciscocmd-cosi/issues/2